WID-51: move function for request blob to handler.ts
